Finalizacion de estadisticas

+++ sgestor.h
@@ -10,13 +10,12 @@
 struct SGestor
 {
     int numUsuarios;
-    int** relaciones;
+    int **relaciones;
     char modo;
-    float tiempo;
+    int tiempo;
     char pipeNom[20];
     struct SCliente clientes[80];
     struct SMensaje mensajes[20];
-
 };
 
 #endif // SGESTOR__H
